# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Plan Holiday",
        "description": "API Plan Holiday",
        "contact": {
            "email": "pe-ronchini@hotmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Register"
                ],
                "summary": "Register a new user",
                "operationId": "register",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User registered successfully"
                    },
                    "400": {
                        "description": "Invalid request data"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Login"
                ],
                "summary": "Login user",
                "operationId": "login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User logged in successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "Invalid request data"
                    }
                }
            }
        },
        "/generatePdfPlanHoliday/{id}": {
            "get": {
                "tags": [
                    "Plan Holiday"
                ],
                "summary": "Generate PDF for plan holiday",
                "operationId": "736205d8b8c18711d1ae4f41b7815ef2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Plan holiday ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "pdfUrl": {
                                            "description": "URL of the generated PDF",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Plan holiday not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": {
                    "passport": {
                        "oauth2": []
                    }
                }
            }
        },
        "/api/showPlanHoliday": {
            "get": {
                "tags": [
                    "Plan Holiday"
                ],
                "summary": "Get all plan holidays",
                "operationId": "974acf49171123b9325f9fbe6809ccd0",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": {
                    "passport": {
                        "oauth2": []
                    }
                }
            }
        },
        "/api/storePlanHoliday": {
            "post": {
                "tags": [
                    "Plan Holiday"
                ],
                "summary": "Store a new plan holiday",
                "operationId": "06a1602fdcea40a3cc1c12e24d10b898",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "New Holiday"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "A new holiday"
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-05-01"
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "New York"
                                    },
                                    "participants": {
                                        "type": "string",
                                        "example": "John Doe, Jane Doe"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "New Holiday"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "A new holiday"
                                        },
                                        "date": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2023-05-01"
                                        },
                                        "location": {
                                            "type": "string",
                                            "example": "New York"
                                        },
                                        "participants": {
                                            "type": "string",
                                            "example": "John Doe, Jane Doe"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "datetime",
                                            "example": "2023-05-01T00:00:00.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "datetime",
                                            "example": "2023-05-01T00:00:00.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    }
                },
                "security": {
                    "passport": {
                        "oauth2": []
                    }
                }
            }
        },
        "/api/showPlanHoliday/{id}": {
            "get": {
                "tags": [
                    "Plan Holiday"
                ],
                "summary": "Get a plan holiday by ID",
                "operationId": "07403c97ac65edf89a45fd22197f69dc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "New Holiday"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "A new holiday"
                                        },
                                        "date": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2023-05-01"
                                        },
                                        "location": {
                                            "type": "string",
                                            "example": "New York"
                                        },
                                        "participants": {
                                            "type": "string",
                                            "example": "John Doe, Jane Doe"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "datetime",
                                            "example": "2023-05-01T00:00:00.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "datetime",
                                            "example": "2023-05-01T00:00:00.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Plan holiday not found"
                    }
                },
                "security": {
                    "passport": {
                        "oauth2": []
                    }
                }
            }
        },
        "/api/updatePlanHoliday/{id}": {
            "put": {
                "tags": [
                    "Plan Holiday"
                ],
                "summary": "Update a plan holiday by ID",
                "operationId": "cdfaa4c039453ecc8b58adf182fc5b3c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Updated Holiday"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "An updated holiday"
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-05-01"
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "New York"
                                    },
                                    "participants": {
                                        "type": "string",
                                        "example": "John Doe, Jane Doe"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Updated Holiday"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "An updated holiday"
                                        },
                                        "date": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2023-05-01"
                                        },
                                        "location": {
                                            "type": "string",
                                            "example": "New York"
                                        },
                                        "participants": {
                                            "type": "string",
                                            "example": "John Doe, Jane Doe"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "datetime",
                                            "example": "2023-05-01T00:00:00.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "datetime",
                                            "example": "2023-05-01T00:00:00.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Plan holiday not found"
                    },
                    "400": {
                        "description": "Validation error"
                    }
                },
                "security": {
                    "passport": {
                        "oauth2": []
                    }
                }
            }
        },
        "/api/deletePlanHoliday/{id}": {
            "delete": {
                "tags": [
                    "Plan Holiday"
                ],
                "summary": "Delete a plan holiday by ID",
                "operationId": "e787305c951271ab3636d170946f5c66",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Plan holiday deleted"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Plan holiday not found"
                    }
                },
                "security": {
                    "passport": {
                        "oauth2": []
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "passport": {
                "type": "apiKey",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://localhost/oauth/token",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    }
}